devkit-max/requirements.txt
mcp>=1.0.0
fastapi>=0.104.0
python-dateutil>=2.8.2
PyJWT>=2.8.0
httpx>=0.25.0
pydantic>=2.4.0
